+++ changelog.d/20241008_141637_sirosen_remove_testing_module.rst
@@ -1,0 +1,5 @@
+Removed
+-------
+
+- ``globus_action_provider_tools.testing`` has been removed. Users who were
+  relying on these components should make use of their own fixtures and mocks.
